fix(issue): move the neutral Badge model out of ob1; drop OB1 keys for OB3
The version-neutral badge config model (Badge, BadgeImgType) lived in the legacy
ob1 package, so every modern module imported the shared model out of the
deprecated tree, and Badge.create_from_conf unconditionally required the OB1
fields verify_key and criteria — issuing an OB3 badge without them blew up with a
raw KeyError('verify_key') straight through issue_from_conf, whose contract
promises IssuanceError.

- Move Badge and BadgeImgType to openbadgeslib.badge_model; ob1.badge re-exports
  them (via __all__), so ob1.badge and the openbadgeslib.badge shim keep working.
- create_from_conf reads verify_key/criteria optionally (they are OB1-only), so
  an OB3-only config issues cleanly; the print + bare `raise IOError` on a
  missing image becomes a typed ErrorParsingFile, and the unreachable
  `return None` is gone.
- issue_from_conf / issue_batch_from_conf wrap create_from_conf so a missing
  required key or an unreadable key/image surfaces as IssuanceError.
- Point issue.py, openbadges_signer and mail at badge_model directly; the OB1
  urls_has_problems download mock moves with the model (badge_model), while the
  OB1 verify path keeps mocking ob1.badge.

Closes #223
