+++ openbadgeslib/ob1/badge.py
@@ -21,17 +21,24 @@
         License along with this library.
 """
 
-import os
 from enum import Enum
 from typing import Any, Optional, Union, cast
 
-from ..keys import KeyType, detect_key_type
+from ..keys import detect_key_type
 from ..errors import (BadgeImgFormatUnsupported, AssertionFormatIncorrect,
-                      ErrorParsingFile, UnknownKeyType, PrivateKeyReadError,
-                      PublicKeyReadError)
+                      ErrorParsingFile)
 from .._jws import utils as jws_utils
 from ..util import hash_email, download_file
 from .. import baking
+# The version-neutral model lives in badge_model now; re-exported here (via
+# __all__) so ob1.badge and the openbadgeslib.badge shim keep exposing them.
+from ..badge_model import Badge, BadgeImgType
+
+__all__ = [
+    'BadgeStatus', 'BadgeImgType', 'BadgeType',
+    'Assertion', 'Badge', 'BadgeSigned',
+    'extract_svg_assertion', 'extract_png_assertion',
+]
 
 
 class BadgeStatus(Enum):
@@ -41,11 +48,6 @@ class BadgeStatus(Enum):
     REVOKED = 4
     IDENTITY_ERROR = 5
     NONE = 6  # unset sentinel (VerifyInfo default); never returned by a real check
-
-
-class BadgeImgType(Enum):
-    SVG = 0
-    PNG = 1
 
 
 class BadgeType(Enum):
@@ -88,152 +90,6 @@ class Assertion():
 
     def __str__(self) -> str:
         return 'Header: %s\nBody: %s\nSignature: %s' % (self.header, self.body, self.signature)
-
-
-class Badge():
-    def __init__(self, ini_name: Optional[str] = None, name: Optional[str] = None,
-                 description: Optional[str] = None, image_type: Optional[BadgeImgType] = None,
-                 image: Optional[bytes] = None, image_url: Optional[str] = None,
-                 criteria_url: Optional[str] = None, json_url: Optional[str] = None,
-                 verify_key_url: Optional[str] = None, key_type: Optional[KeyType] = None,
-                 privkey_pem: Optional[Union[str, bytes]] = None,
-                 pubkey_pem: Optional[Union[str, bytes]] = None) -> None:
-
-        self.ini_name = ini_name
-        self.name = name
-        self.description = description
-        self.image_type = image_type
-        self.image = image                  # Binary contents of image file
-        self.image_url = image_url
-        self.criteria_url = criteria_url
-        self.json_url = json_url
-        self.verify_key_url = verify_key_url
-        self.key_type = key_type
-        self.privkey_pem = privkey_pem
-        self.pubkey_pem = pubkey_pem
-
-        # Default to no key material so a Badge built without PEMs exposes
-        # priv_key/pub_key as None rather than leaving the attributes undefined
-        # (which turns a later read in the signer/verifier into a confusing raw
-        # AttributeError instead of a clean "no key material" case).
-        self.priv_key: Any = None
-        self.pub_key: Any = None
-
-        # Load whatever key material was supplied. RSA/ECC used to go through
-        # pycryptodome/python-ecdsa; all three families now share cryptography's
-        # unambiguous PEM loaders (the #167 port), producing cryptography key
-        # objects that the JWS layer serialises via keys.key_to_pem.
-        if self.key_type in (KeyType.RSA, KeyType.ECC, KeyType.ED25519):
-            from cryptography.hazmat.primitives.serialization import (
-                load_pem_private_key, load_pem_public_key)
-            label = self.key_type.name
-            if self.pubkey_pem:
-                try:
-                    pub_pem = self.pubkey_pem.encode('utf-8') \
-                        if isinstance(self.pubkey_pem, str) else self.pubkey_pem
-                    self.pub_key = load_pem_public_key(pub_pem)
-                except Exception as exc:
-                    raise PublicKeyReadError(
-                        'Unable to read %s public key: %s' % (label, exc)) from exc
-            if self.privkey_pem:
-                try:
-                    priv_pem = self.privkey_pem.encode('utf-8') \
-                        if isinstance(self.privkey_pem, str) else self.privkey_pem
-                    self.priv_key = load_pem_private_key(priv_pem, password=None)
-                except Exception as exc:
-                    raise PrivateKeyReadError(
-                        'Unable to read %s private key: %s' % (label, exc)) from exc
-        elif self.key_type is not None:
-            # key_type=None is a valid "no key material yet" state; any other
-            # value is an unsupported key type and must fail loudly.
-            raise UnknownKeyType('Unsupported key type: %r' % (self.key_type,))
-
-    @staticmethod
-    def create_from_conf(conf: Any, badge: str) -> 'Badge':
-        """ Create a Badge Object reading params from config.ini """
-
-        """ Keys """
-        try:
-            with open(conf[badge]['private_key'], 'rb') as key:
-                privkey_pem = key.read()
-        except OSError as exc:
-            raise PrivateKeyReadError(
-                'Unable to read private key file %s: %s' % (conf[badge]['private_key'], exc)) from exc
-
-        try:
-            with open(conf[badge]['public_key'], 'rb') as key:
-                pubkey_pem = key.read()
-        except OSError as exc:
-            raise PublicKeyReadError(
-                'Unable to read public key file %s: %s' % (conf[badge]['public_key'], exc)) from exc
-
-        key_type = detect_key_type(pubkey_pem)
-
-        """ Image """
-        img_path = os.path.join(conf['paths']['base_image'], conf[badge]['local_image'])
-
-        if not os.path.isfile(img_path):
-            print('Badge file %s NOT exists.' % img_path)
-            raise IOError
-
-        with open(img_path, 'rb') as file:
-            img_content = file.read()
-
-        if img_path.lower().endswith('.svg'):
-            img_type = BadgeImgType.SVG
-        elif img_path.lower().endswith('.png'):
-            img_type = BadgeImgType.PNG
-        else:
-            raise BadgeImgFormatUnsupported('The image format for %s is not supported' % badge)
-
-        """ Object Creation """
-        return Badge(ini_name=badge,
-                     name=conf[badge]['name'],
-                     description=conf[badge]['description'],
-                     image_type=img_type,
-                     image=img_content,
-                     image_url=conf[badge]['image'],
-                     criteria_url=conf[badge]['criteria'],
-                     json_url=conf[badge]['badge'],
-                     verify_key_url=conf[badge]['verify_key'],
-                     key_type=key_type,
-                     privkey_pem=privkey_pem,
-                     pubkey_pem=pubkey_pem)
-
-        return None
-
-    def __str__(self) -> str:
-        return (
-            'INI Name: %s\nName: %s\nDescription: %s\nImage Type: %s\n'
-            'Image Url: %s\nKey Type: %s\nVerify Key: %s\nJSON Url: %s\n'
-            % (self.ini_name, self.name, self.description, self.image_type,
-               self.image_url, self.key_type, self.verify_key_url, self.json_url)
-        )
-
-    def urls_has_problems(self) -> bool:
-        """ Check if urls in Badge are corrects and online """
-
-        error = False
-        checks = [
-            (self.image_url, 'OpenBadge Image'),
-            (self.criteria_url, 'OpenBadge Criteria'),
-            (self.json_url, 'OpenBadge JSon'),
-            (self.verify_key_url, 'OpenBadge Verify key'),
-        ]
-
-        # 'data' is reset for every URL so a previous success can never mask a
-        # later download failure.
-        for url, label in checks:
-            data = None
-            try:
-                data = download_file(url) if url else None
-            except Exception:
-                pass
-            if not data:
-                print('[!] %s at config file is pointing to a wrong url: %s' % (label, url))
-                error = True
-
-        return error
 
 
 class BadgeSigned():
